+++ databricks-default-cluster-policies/variables.tf
@@ -29,3 +29,14 @@ variable "personal_compute_pool_ids" {
   type = list(string)
   default = []
 }
+
+variable "policy_map" {
+  description = "Map of policy names to groups"
+  type = list(map(list(string)))
+}
+
+variable "power_user_group_name" {
+  description = "Name of the power user group"
+  type = string
+  default = "Power Users"
+}